Basically adds a particle system
It was not the intent.. initially, but I'm too enchanted by the
simplicity of it all to take it out/disable it.

Also fixed the random velocity applied to instances of the Ray,
and also added Ray::reflect, although the first couple of tries
had some bugs in it (naturally). The biggest one was the `elif`
block's reflection - multiplying a value by 1 does not reflect
the value.. it keeps it the same. D'oh!