Does some things
* Some tweaks on the environments, to send updates to entities within. Not
quite happy with it just yet, but don't want to deal with it right now.

* TIME! Oi, this was... a thing. Time now is reported I believe in seconds,
but there's an early measurement system now available, but not yet applied,
so that if a timer reports a time in a particular unit, it can be converted
back to the base unit, ready to apply to the environment's desired timing
interval. Only tested in the conveniently-named file, testing.py.

* Demos moved to the root directory! Now, if you're using an integrated
development environment, you can, pending configuration within, press the
"run"/"debug" button with that file opened, or invoke it directly using a
`python <demoFileName>` call. This al;so obsoleted the need for the main
launcher.py file.. so it's gone.

I was also messing around with using `c` - the constant for the speed of
light in a vacuum, but the rays kept doing weird (albeit fun!) things, so
I undid that, but it was what caused me to write an initial units ane
measurements system. The one included here is similar, but is a second
attempt, and one I like enough to push publicly. d^_^b