Adds a sorting demo with.. I think bubble sort?
Could be wrong, but just wanted to create it, based on a link in a
Discord guild when I posted a picture of the updated emitter demo. I
was informed it reminded them a bit of some older sorting videos, so
I reviewed a few, and saw one that looked like it would be easy to
replicate, to some degree. So I did.